SONAR-6822 Move search action tests to a dedicated test class
