+++ server/sonar-server/src/test/resources/org/sonar/server/rule/ws/SearchActionMediumTest/search_active_rules.json
