+++ server/sonar-server/src/test/resources/org/sonar/server/rule/ws/SearchActionMediumTest/search_debt_rules_sticky.json
