+++ server/sonar-server/src/test/resources/org/sonar/server/rule/ws/SearchActionMediumTest/severities_facet.json
