+++ server/sonar-server/src/test/resources/org/sonar/server/rule/ws/SearchActionMediumTest/statuses_facet.json
